test(server): move the drain-defer regression next to cloud-provider-sync as a bun test
evals/specs must cross a product boundary; this test wires EnginePool and
startServer directly, so the boundary ratchet rejects it there. Same three
claims (boundedRun, drainUntouched, retryLands) as bun tests.
